Feat/5 store dto create (#77)
store dto create
    - request 
         - StoreCreateRequest
         - StoreUpdateRequest
    - response
         - StoreListResponse
         - StoreResponse

---------

+++ src/main/java/com/example/Spot/store/domain/repository/StoreRepository.java
@@ -41,7 +41,7 @@ public interface StoreRepository extends JpaRepository<StoreEntity, UUID> {
             "WHERE u.id = :userId " +
             "AND u.role = ('OWNER', 'CHEF') " +
             "AND s.isDeleted = false")
-    List<StoreEntity> findAllByOwnerId(@Param("userId") Long userId);
+    List<StoreEntity> findAllByOwnerId(@Param("userId") Integer userId);
 
     // category-repo
     // 삭제되지 않은 가게 전체 조회
